% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abundance.R
\name{calc_abundance}
\alias{calc_abundance}
\title{Calculate abundance
Using number of individuals. Can be sex- and/or age-specific.}
\usage{
calc_abundance(N_prev, M, F_ret, F_inc)
}
\arguments{
\item{N_prev}{Abundance (number of individuals) from previous time step}

\item{M}{Natural mortality}

\item{F_ret}{Retained fishing mortality}

\item{F_inc}{Incidental fishing mortality}
}
\value{
Abundance (number of individuals) in the next time step
}
\description{
Calculate abundance
Using number of individuals. Can be sex- and/or age-specific.
}
