% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_BevHolt.R
\name{calc_BevHolt}
\alias{calc_BevHolt}
\title{Calculate Beverton-Holt recruitment
Assumes females will only mate with males older than them. Assumes oldest females will reproduce with oldest males.
Follows parametrisation from Brooks and Powers 2007}
\usage{
calc_BevHolt(h, R0, phi0, E)
}
\arguments{
\item{h}{steepness}

\item{R0}{virgin recruitment}

\item{phi0}{virgin level of eggs per recruit}

\item{E}{number of eggs; the number at time 0 of stage 1}
}
\description{
Calculate Beverton-Holt recruitment
Assumes females will only mate with males older than them. Assumes oldest females will reproduce with oldest males.
Follows parametrisation from Brooks and Powers 2007
}
