% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_catch.R
\name{calc_catch}
\alias{calc_catch}
\title{Calculate catch
Based on Baranov catch equation. Not totally sure what units are; not based on abundance?
Can be sex- and/or age-specific.}
\usage{
calc_catch(N, M, F_ret, F_inc)
}
\arguments{
\item{N}{Abundance of exploitable population}

\item{M}{Natural mortality}

\item{F_ret}{Retained fishing mortality}

\item{F_inc}{Incidental fishing mortality}
}
\value{
Catch for that time step. Not totally sure what the units are.
}
\description{
Calculate catch
Based on Baranov catch equation. Not totally sure what units are; not based on abundance?
Can be sex- and/or age-specific.
}
