% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eggprod.R
\name{calc_eggprod}
\alias{calc_eggprod}
\title{Calculate egg production
Assumes females will only mate with males older than them. Assumes oldest females will reproduce with oldest males.}
\usage{
calc_eggprod(N_female, N_male, gamma_par, delta, eta)
}
\arguments{
\item{N_female}{Vector of female abundance by age}

\item{N_male}{Vector of male abundance by age}

\item{gamma_par}{Vector of fecundity-at-age}

\item{delta}{Vector of proportion of mature females at age a}

\item{eta}{Half saturation constant}
}
\value{
Incidental fishing mortality
}
\description{
Calculate egg production
Assumes females will only mate with males older than them. Assumes oldest females will reproduce with oldest males.
}
