% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_F_inc.R
\name{calc_F_inc}
\alias{calc_F_inc}
\title{Calculate incidental fishing mortality
Can be sex-, age-, and time-specific.}
\usage{
calc_F_inc(u, omega, sigma)
}
\arguments{
\item{u}{Probability of being captured}

\item{omega}{Probability of being retained}

\item{sigma}{Probability of incidental death}
}
\value{
Incidental fishing mortality
}
\description{
Calculate incidental fishing mortality
Can be sex-, age-, and time-specific.
}
