% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_F_ret.R
\name{calc_F_ret}
\alias{calc_F_ret}
\title{Calculate retained fishing mortality
Can be sex- and/or age-specific.}
\usage{
calc_F_ret(u, omega)
}
\arguments{
\item{u}{Probability of being captured}

\item{omega}{Probability of being retained}
}
\value{
Retained fishing mortality
}
\description{
Calculate retained fishing mortality
Can be sex- and/or age-specific.
}
