% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_u.R
\name{calc_u}
\alias{calc_u}
\title{Calculate probability of capture
Can be sex-, age-, and time-specific.}
\usage{
calc_u(q, Effort)
}
\arguments{
\item{q}{Catchability}

\item{Effort}{Total effort at time step}
}
\value{
Probability of capture
}
\description{
Calculate probability of capture
Can be sex-, age-, and time-specific.
}
