% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_week.R
\name{calc_week}
\alias{calc_week}
\title{Calculate week duration between two dates}
\usage{
calc_week(curdt, stdt = "November 1, 1999")
}
\arguments{
\item{curdt}{Date to calculate to in string format, e.g., "August 27 2019"}

\item{stdt}{Date to calculate from in string format. Default is "November 1, 1999"}
}
\value{
Difference in two dates in weeks
}
\description{
Calculate week duration between two dates
}
